Add Backpack.tf pricing integration with conversion support
