+++ templates/_user.html
@@ -23,7 +23,7 @@
     {% endif %}
     <div class="items">
       {% for item in user.items %}
-        <div class="item-card" style="border-color: {{ item.quality_color }};">
+        <div class="item-card{% if item.unknown %} unknown{% endif %}" style="border-color: {{ item.quality_color }};{% if item.unknown %} opacity:0.5;{% endif %}">
           {% if item.final_url %}
             <img src="{{ item.final_url }}" alt="{{ item.name }}" width="32" height="32">
           {% else %}
