+++ tests/test_pricing_service.py
@@ -1,11 +1,11 @@
-import utils.price_fetcher as pf
+import services.pricing_service as ps
 
 
 def test_price_cache_fetch(tmp_path, monkeypatch):
     price_file = tmp_path / "price_cache.json"
     curr_file = tmp_path / "currency_rates.json"
-    monkeypatch.setattr(pf, "PRICE_CACHE_FILE", price_file)
-    monkeypatch.setattr(pf, "CURRENCY_FILE", curr_file)
+    monkeypatch.setattr(ps, "PRICE_CACHE_FILE", price_file)
+    monkeypatch.setattr(ps, "CURRENCY_FILE", curr_file)
 
     class DummyResp:
         def __init__(self, payload):
@@ -40,11 +40,11 @@ def test_price_cache_fetch(tmp_path, monkeypatch):
             return DummyResp(prices_payload)
         return DummyResp(currencies_payload)
 
-    monkeypatch.setattr(pf.requests, "get", fake_get)
+    monkeypatch.setattr(ps.requests, "get", fake_get)
     monkeypatch.setenv("BACKPACK_API_KEY", "x")
 
-    prices = pf.ensure_price_cache()
-    rate = pf.ensure_currency_rates()
+    prices = ps.ensure_price_cache()
+    rate = ps.ensure_currency_rates()
 
     assert prices["30035;6"]["value"] == 5100
     assert rate == 51.0
@@ -53,5 +53,5 @@ def test_price_cache_fetch(tmp_path, monkeypatch):
 
 
 def test_format_price():
-    result = pf.format_price(5150, 5100)
+    result = ps.format_price(5150, 5100)
     assert "key" in result and "0.50" in result
